float instead of double

+++ q1/ConvexHull.cpp
@@ -3,7 +3,7 @@
 #include <algorithm>
 
 // Cross product
-static double cross(const Point& O, const Point& A, const Point& B) {
+static float cross(const Point& O, const Point& A, const Point& B) {
     return (A.x - O.x) * (B.y - O.y) - (A.y - O.y) * (B.x - O.x);
 }
