+++ q1/Polygon.cpp
@@ -2,8 +2,8 @@
 #include "Polygon.hpp"
 #include <cmath>
 
-double polygon_area(const std::vector<Point>& poly) {
-    double area = 0;
+float polygon_area(const std::vector<Point>& poly) {
+    float area = 0;
     int n = poly.size();
     for (int i = 0; i < n; ++i) {
         area += poly[i].x * poly[(i+1)%n].y;
